chore: update tests and scripts

+++ test/utils/Constants.sol
@@ -33,9 +33,9 @@ abstract contract Constants {
     bytes32 internal constant MINT_FEATURE_FLAG = ProtocolConstants.MINT_FEATURE_FLAG;
 
     /// @notice Margin collateral types configuration constants.
-    uint256 internal constant USDC_DEPOSIT_CAP = 50_000_000_000e18;
-    uint256 internal constant WSTETH_DEPOSIT_CAP = 1_000_000e18;
-    uint256 internal constant ZRSUSD_DEPOSIT_CAP = 50_000_000_000e18;
+    uint248 internal constant USDC_DEPOSIT_CAP = 50_000_000_000e18;
+    uint248 internal constant WSTETH_DEPOSIT_CAP = 1_000_000e18;
+    uint248 internal constant ZRSUSD_DEPOSIT_CAP = 50_000_000_000e18;
 
     /// @notice ETH/USD market configuration constants.
     uint128 internal constant ETH_USD_MARKET_ID = 1;
